# Orthogonal defaulting and new pruning algorithm