--- a/defaulter_test.go
+++ b/defaulter_test.go
@@ -30,23 +30,13 @@ import (
 var defaulterFixturesPath = filepath.Join("fixtures", "defaulting")
 
 func TestDefaulter(t *testing.T) {
-	fname := filepath.Join(defaulterFixturesPath, "schema.json")
-	b, err := ioutil.ReadFile(fname)
+	schema, err := defaulterFixture()
 	assert.NoError(t, err)
-	var schema spec.Schema
-	assert.NoError(t, json.Unmarshal(b, &schema))
 
-	err = spec.ExpandSchema(&schema, nil, nil /*new(noopResCache)*/)
-	assert.NoError(t, err, fname+" should expand cleanly")
-
-	validator := NewSchemaValidator(&schema, nil, "", strfmt.Default)
-	x := map[string]interface{}{
-		"nested": map[string]interface{}{},
-		"all":    map[string]interface{}{},
-		"any":    map[string]interface{}{},
-		"one":    map[string]interface{}{},
-	}
+	validator := NewSchemaValidator(schema, nil, "", strfmt.Default)
+	x := defaulterFixtureInput()
 	t.Logf("Before: %v", x)
+
 	r := validator.Validate(x)
 	assert.False(t, r.HasErrors(), fmt.Sprintf("unexpected validation error: %v", r.AsError()))
 
@@ -98,4 +88,42 @@ func TestDefaulterSimple(t *testing.T) {
 	}`), &expected)
 	assert.NoError(t, err)
 	assert.Equal(t, expected, x)
+}
+
+func BenchmarkDefaulting(b *testing.B) {
+	b.ReportAllocs()
+	
+	schema, err := defaulterFixture()
+	assert.NoError(b, err)
+
+	for n := 0; n < b.N; n++ {
+		validator := NewSchemaValidator(schema, nil, "", strfmt.Default)
+		x := defaulterFixtureInput()
+		r := validator.Validate(x)
+		assert.False(b, r.HasErrors(), fmt.Sprintf("unexpected validation error: %v", r.AsError()))
+		r.ApplyDefaults()
+	}
+}
+
+func defaulterFixtureInput() map[string]interface{} {
+	return map[string]interface{}{
+		"nested": map[string]interface{}{},
+		"all":    map[string]interface{}{},
+		"any":    map[string]interface{}{},
+		"one":    map[string]interface{}{},
+	}
+}
+
+func defaulterFixture() (*spec.Schema, error) {
+	fname := filepath.Join(defaulterFixturesPath, "schema.json")
+	b, err := ioutil.ReadFile(fname)
+	if err != nil {
+		return nil, err
+	}
+	var schema spec.Schema
+	if err := json.Unmarshal(b, &schema); err != nil {
+		return nil, err
+	}
+
+	return &schema, spec.ExpandSchema(&schema, nil, nil /*new(noopResCache)*/)
 }