--- a/schema.go
+++ b/schema.go
@@ -98,9 +98,12 @@ func (s *SchemaValidator) Applies(source interface{}, kind reflect.Kind) bool {
 
 // Validate validates the data against the schema
 func (s *SchemaValidator) Validate(data interface{}) *Result {
-	result := new(Result)
+	result := &Result{data: data}
 	if s == nil {
 		return result
+	}
+	if s.Schema != nil {
+		result.addRootObjectSchemata(s.Schema)
 	}
 
 	if data == nil {
@@ -162,6 +165,7 @@ func (s *SchemaValidator) Validate(data interface{}) *Result {
 		result.Inc()
 	}
 	result.Inc()
+
 	return result
 }