… fixtures/pruning/schema.json