This PR allows to implement different algorithms orthogonally to the main
validation recursion. The Result object tracks the validating schemata for
each node in the input JSON. This means that these algorithms do not
have to care about quantors or the different ways to specify properties.

This means:

- defaulting becomes a trival 12 liner:
```golang
// ApplyDefaults applies defaults to data.
func ApplyDefaults(r *validate.Result) {
	fieldSchemata := r.FieldSchemata()
	for key, schemata := range fieldSchemata {
	LookForDefaultingScheme:
		for _, s := range schemata {
			if s.Default != nil {
				key.Object()[key.Field()] = s.Default
				break LookForDefaultingScheme
			}
		}
	}
}
```
- pruning of unspecified fields is implementable in a similar way. It's added as another example.

Running defaulting benchmarks 30000 leads to:
```
Before PR:
30000	     44573 ns/op	   54581 B/op	     499 allocs/op
30000	     43208 ns/op	   54581 B/op	     499 allocs/op
30000	     39850 ns/op	   54582 B/op	     499 allocs/op
30000	     39427 ns/op	   54582 B/op	     499 allocs/op
```
After PR:
```
30000	     48023 ns/op	   66088 B/op	     497 allocs/op
30000	     46331 ns/op	   66086 B/op	     497 allocs/op
```

In other words: runtime is 15% higher, memory is 20% higher. If this is a problem, we could add a second `ValidateWithSchemaTracking()` func which does the extra processing and returns a result that has the necessary getters.

TODO:
- [x] benchmark cpu and memory/allocations: