@fredbi 2018-03-07
I like the idea of metadata in the result structure. 
That is something I find useful to spew more accurate messages on _where_ the validation error did occur.
But I am not sure this is what you had in mind.
Your PR is lagging master for several months I think.