@sttts 2018-03-07
> Your PR is lagging master for several months I think.

Rebased already.

Will push the second algorithm soon to show what this is all about.