@sttts 2018-03-07
@fredbi compare the last commit. This is the kind of trivial algorithm I have in mind, without complicating the main validation recursion.