@sttts 2018-04-13
@fredbi please take another look. This is ready for review.