@casualjim 2018-04-15
this is causing issues downstream, I'm going to revert this PR.. Will try to add the failing cases to this CI build so we discover this earlier.