@casualjim 2018-04-15
After applying this PR, the types in slice properties are no longer []interface{}

Here's a job that fails:
https://travis-ci.org/go-openapi/runtime/jobs/366714797


just running the tests in go-openapi/runtime shows the issues.