@sttts 2018-04-15
@casualjim will look into that tomorrow.

Some integration CI would be nice. How do you notice those problems now? Just merging and then seeing regression on the next downstream runs?