@casualjim 2018-04-15
yeah basically.
I would like to add the following to .travis.yml

```yaml
- go get -u github.com/go-openapi/runtime
- cd $GOPATH/src/github.com/go-openapi/runtime
- go test ./...
```

to run after the tests of the validate repo have finished