@sttts 2018-04-15
I took a quick look at the issue: my change in the `schemaSliceValidator` adds the requirement that the input `data` is a JSON like `[]interface{}` slice, but here it is not: it's a `[]middleware.friend`. We used reflection before my change to work around this. 

I noticed that the `objectValidator` in contrast has a hard requirement on  `map[string]interface{}` which sounds correct.

Am I right that the `data` used as validator input must be JSON? I.e. not an arbitrary Go type? If yes, the runtime middleware test broken it seems.