@casualjim 2018-04-15
yes you're right, the input should be json for the json schema validators