@sttts 2018-04-16
Working on the go-openapi/runtime fix now.